+++ NEWS.md
@@ -1,7 +1,7 @@
 # hermes 0.1.1.9000
 
 ### Miscellaneous
-* A new utility function `cut_quantile` replaces the `tern::cut_quantile_bins` in `draw_barplot`.
+* New utility function `cut_quantile` for cutting a numeric vector into quantiles.
 
 # hermes 0.1.1